+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add GH Actions release-* and publish-*
+
 ## 1.1.0 - 2021-06-20
 ### Added
 - Initial version